# coverup-impl/rest-openapi.yaml
info:
  description: ""
  license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  title: "Coverup"
  version: v1.0
openapi: 3.0.1
components:
  schemas:
    CoverupObject:
      description: A service to coverup data within your instance. Useful for using Production data in UAT or DEV. CoverupObject makes no sense. Maybe change to StatusObject.
      properties:
        name:
          description: The entity name.
          type: string
        id:
          description: The entity ID.
          type: integer
      type: object
    SuperheroObject:
      description: A service to coverup data within your instance. Useful for using Production data in UAT or DEV. CoverupObject makes no sense. Maybe change to StatusObject.
      properties:
        name:
          description: Hero.
          type: string
        power:
          description: Power.
          type: integer
      type: object
paths:
  "/alive":
    get:
      responses:
        200:
          description: "Quick test to see it's alive with dummy response. No worries."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuperheroObject"
            application/xml:
              schema:
                $ref: "#/components/schemas/SuperheroObject"
      tags: ["Coverup"]
  "/{confirmationId}":
    post:
      parameters:
        - in: path
          name: confirmationId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: "Use confirmationId as defined in portal-ext.properties"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverupObject"
            application/xml:
              schema:
                $ref: "#/components/schemas/CoverupObject"
      tags: ["Coverup"]
  "/Users/{confirmationId}":
    post:
      parameters:
        - in: path
          name: confirmationId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: "Use confirmationId as defined in portal-ext.properties"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverupObject"
            application/xml:
              schema:
                $ref: "#/components/schemas/CoverupObject"
      tags: ["Coverup"]